extract philosopher utils
